Render tesselations as simple angle paths
A tesselation can be represented as a tree of angles. Each node contains
an angle relative to its parent. Each child is a connected vertex.

My current implementation is 1) incorrect (the angles are off so the
whole graph is heavily disjoined) and 2) exponential (there is no
bounding so each vertex is represented by an exponential number of
nodes.

So ... it needs amendments.